+++ src/MachineLearningServices/MachineLearningServices.Autorest/custom/New-AzMLWorkspaceCustomModelJobInputObject.ps1
@@ -16,38 +16,39 @@
 
 <#
 .Synopsis
-Create an in-memory object for MLFlowModelJobOutput.
+Create an in-memory object for CustomModelJobInput.
 .Description
-Create an in-memory object for MLFlowModelJobOutput.
+Create an in-memory object for CustomModelJobInput.
 
 .Outputs
-Microsoft.Azure.PowerShell.Cmdlets.MachineLearningServices.Models.Api20240401.MLFlowModelJobOutput
+Microsoft.Azure.PowerShell.Cmdlets.MachineLearningServices.Models.Api20240401.CustomModelJobInput
 .Link
-https://learn.microsoft.com/powershell/module/Az.MachineLearningServices/new-AzMLWorkspaceMLFlowModelJobOutputObject
+https://learn.microsoft.com/powershell/module/Az.MachineLearningServices/new-AzMLWorkspaceCustomModelJobInputObject
 #>
-function New-AzMLWorkspaceMLFlowModelJobOutputObject {
-    [OutputType('Microsoft.Azure.PowerShell.Cmdlets.MachineLearningServices.Models.Api20240401.MLFlowModelJobOutput')]
+function New-AzMLWorkspaceCustomModelJobInputObject {
+    [OutputType('Microsoft.Azure.PowerShell.Cmdlets.MachineLearningServices.Models.Api20240401.CustomModelJobInput')]
     [CmdletBinding(PositionalBinding=$false)]
     Param(
 
-        [Parameter(HelpMessage="Output Asset Delivery Mode.")]
-        [ArgumentCompleter([Microsoft.Azure.PowerShell.Cmdlets.MachineLearningServices.Support.OutputDeliveryMode])]
-        [Microsoft.Azure.PowerShell.Cmdlets.MachineLearningServices.Support.OutputDeliveryMode]
+        [Parameter(HelpMessage="Input Asset Delivery Mode.")]
+        [ArgumentCompleter([Microsoft.Azure.PowerShell.Cmdlets.MachineLearningServices.Support.InputDeliveryMode])]
+        [Microsoft.Azure.PowerShell.Cmdlets.MachineLearningServices.Support.InputDeliveryMode]
         $Mode,
-        [Parameter(HelpMessage="Output Asset URI.")]
+        [Parameter(Mandatory, HelpMessage="[Required] Input Asset URI.")]
         [string]
         $Uri,
-        [Parameter(HelpMessage="Description for the output.")]
+        [Parameter(HelpMessage="Description for the input.")]
         [string]
         $Description,
+        [Microsoft.Azure.PowerShell.Cmdlets.MachineLearningServices.Runtime.ParameterBreakingChangeAttribute("Type", "16.0.0", "1.3.0", "2026/05")]
         [Parameter(Mandatory, HelpMessage="[Required] Specifies the type of job.")]
-        [ArgumentCompleter([Microsoft.Azure.PowerShell.Cmdlets.MachineLearningServices.Support.JobOutputType])]
-        [Microsoft.Azure.PowerShell.Cmdlets.MachineLearningServices.Support.JobOutputType]
+        [ArgumentCompleter([Microsoft.Azure.PowerShell.Cmdlets.MachineLearningServices.Support.JobInputType])]
+        [Microsoft.Azure.PowerShell.Cmdlets.MachineLearningServices.Support.JobInputType]
         $Type
     )
 
     process {
-        $Object = [Microsoft.Azure.PowerShell.Cmdlets.MachineLearningServices.Models.Api20240401.MLFlowModelJobOutput]::New()
+        $Object = [Microsoft.Azure.PowerShell.Cmdlets.MachineLearningServices.Models.Api20240401.CustomModelJobInput]::New()
 
         if ($PSBoundParameters.ContainsKey('Mode')) {
             $Object.Mode = $Mode